+++ pkg/release/bumpall.go
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	"github.com/blang/semver"
-	tm "github.com/buger/goterm"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/release-operator/v4/api/v1alpha1"
 	"github.com/google/go-github/v74/github"
@@ -44,7 +43,7 @@ type appVersion struct {
 
 // BumpAll takes all apps and components in the `input` release and looks up on github for the latest version of each.
 // If the version is not specified in the `manuallyRequestedComponents` or `manuallyRequestedApps` it will be bumped to the latest version.
-func BumpAll(input v1alpha1.Release, manuallyRequestedComponents []string, manuallyRequestedApps []string, appsToDrop map[string]bool, yes bool, output string, changesOnly bool, requestedOnly bool, k8sMajorVersion uint64) ([]string, []string, error) {
+func BumpAll(input v1alpha1.Release, manuallyRequestedComponents []string, manuallyRequestedApps []string, releaseType string, appsToDrop map[string]bool, requests []Request, yes bool, output string, changesOnly bool, requestedOnly bool, k8sMajorVersion uint64) ([]string, []string, error) {
 	requestedComponents := map[string]componentVersion{}
 	requestedApps := map[string]appVersion{}
 
@@ -77,9 +76,31 @@ func BumpAll(input v1alpha1.Release, manuallyRequestedComponents []string, manua
 				version := componentVersion{}
 
 				if comp.Name == "kubernetes" {
-					version.Version, err = getLatestK8sVersion(k8sMajorVersion)
+					if releaseType == "patch" {
+						// For a patch release, we don't want to automatically bump anything.
+						// The user must manually request a bump for a component.
+						version.Version = comp.Version
+					} else { // major or minor
+						version.Version, err = getLatestK8sVersion(k8sMajorVersion)
+					}
+				} else if comp.Name == "flatcar" {
+					if releaseType == "patch" {
+						version.Version = comp.Version
+					} else { // minor or major
+						version.Version, err = getLatestFlatcarRelease()
+					}
 				} else {
-					version, err = findNewestComponent(comp.Name)
+					var latestVersionString string
+					latestVersionString, err = findNewestComponentVersion(comp.Name)
+					if err == nil {
+						if releaseType == "patch" {
+							// For a patch release, we don't want to automatically bump anything.
+							// The user must manually request a bump for a component.
+							version.Version = comp.Version
+						} else { // major or minor, no restrictions for other components
+							version.Version = latestVersionString
+						}
+					}
 				}
 
 				if err != nil {
@@ -90,6 +111,31 @@ func BumpAll(input v1alpha1.Release, manuallyRequestedComponents []string, manua
 				v.UserRequested = false
 			}
 			if v.Version != comp.Version {
+				// We have a new version, let's check against requests.
+				var constraint semver.Range
+				for _, r := range requests {
+					if r.Name == comp.Name {
+						c, err := semver.ParseRange(r.Version)
+						if err != nil {
+							// Ignore invalid constraints.
+							continue
+						}
+						constraint = c
+						break
+					}
+				}
+
+				if constraint != nil {
+					newV, err := semver.Parse(v.Version)
+					if err != nil {
+						return nil, nil, microerror.Mask(err)
+					}
+					if !constraint(newV) {
+						// Does not meet constraint, reverting to old version.
+						v.Version = comp.Version
+					}
+				}
+
 				components[comp.Name] = v
 			}
 		}
@@ -156,17 +202,49 @@ func BumpAll(input v1alpha1.Release, manuallyRequestedComponents []string, manua
 					v.DependsOn = app.DependsOn
 				}
 			} else {
-				version, err := findNewestApp(app.Name, app.ComponentVersion != "")
-				if err != nil {
-					return nil, nil, microerror.Mask(err)
+				if releaseType == "patch" {
+					v.Version = app.Version
+					v.UpstreamVersion = app.ComponentVersion
+					v.UserRequested = false
+					v.DependsOn = app.DependsOn
+				} else { // major or minor
+					version, err := findNewestApp(app.Name, app.ComponentVersion != "")
+					if err != nil {
+						return nil, nil, microerror.Mask(err)
+					}
+					v.Version = version.Version
+					v.UpstreamVersion = version.UpstreamVersion
+					v.UserRequested = false
+					v.DependsOn = app.DependsOn
 				}
-
-				v.Version = version.Version
-				v.UpstreamVersion = version.UpstreamVersion
-				v.UserRequested = false
-				v.DependsOn = app.DependsOn
 			}
 			if v.Version != app.Version || !slices.Equal(v.DependsOn, app.DependsOn) {
+				// We have a new version, let's check against requests.
+				var constraint semver.Range
+				for _, r := range requests {
+					if r.Name == app.Name {
+						c, err := semver.ParseRange(r.Version)
+						if err != nil {
+							// Ignore invalid constraints.
+							continue
+						}
+						constraint = c
+						break
+					}
+				}
+
+				if constraint != nil {
+					newV, err := semver.Parse(v.Version)
+					if err != nil {
+						return nil, nil, microerror.Mask(err)
+					}
+					if !constraint(newV) {
+						// Does not meet constraint, reverting to old version.
+						v.Version = app.Version
+						v.UpstreamVersion = app.ComponentVersion
+					}
+				}
+
 				apps[app.Name] = v
 			}
 		}
@@ -243,8 +321,6 @@ func BumpAll(input v1alpha1.Release, manuallyRequestedComponents []string, manua
 
 // Just print a table with a list of apps and components with old and new version for easy checking by user.
 func printTable(input v1alpha1.Release, components map[string]componentVersion, apps map[string]appVersion, appsToDrop map[string]bool, output string, changesOnly bool, requestedOnly bool) error {
-	tm.Clear()
-
 	// --- APPS TABLE ---
 	var appRows []table.Row
 	for _, app := range input.Spec.Apps {
@@ -374,9 +450,8 @@ func printTable(input v1alpha1.Release, components map[string]componentVersion, 
 
 	if len(appRows) > 0 {
 		t := table.NewWriter()
-		if output == "text" {
-			t.SetOutputMirror(tm.Output)
-		}
+		t.SetOutputMirror(os.Stdout)
+		t.SetStyle(table.StyleDefault)
 		t.AppendHeader(table.Row{"APP NAME", "CURRENT APP VERSION", "DESIRED APP VERSION", "DEPENDENCIES"})
 		t.AppendSeparator()
 		t.AppendRows(appRows)
@@ -447,9 +522,8 @@ func printTable(input v1alpha1.Release, components map[string]componentVersion, 
 
 	if len(componentRows) > 0 {
 		t := table.NewWriter()
-		if output == "text" {
-			t.SetOutputMirror(tm.Output)
-		}
+		t.SetOutputMirror(os.Stdout)
+		t.SetStyle(table.StyleDefault)
 		t.AppendHeader(table.Row{"COMPONENT NAME", "CURRENT VERSION", "DESIRED VERSION"})
 		t.AppendSeparator()
 		t.AppendRows(componentRows)
@@ -462,7 +536,6 @@ func printTable(input v1alpha1.Release, components map[string]componentVersion, 
 		}
 	}
 
-	tm.Flush()
 	return nil
 }
 
@@ -504,7 +577,7 @@ func findNewestApp(name string, getUpstreamVersion bool) (appVersion, error) {
 	return ret, nil
 }
 
-func findNewestComponent(name string) (componentVersion, error) {
+func findNewestComponentVersion(name string) (string, error) {
 	var err error
 	version := ""
 
@@ -512,28 +585,37 @@ func findNewestComponent(name string) (componentVersion, error) {
 	case "flatcar":
 		version, err = getLatestFlatcarRelease()
 		if err != nil {
-			return componentVersion{}, microerror.Mask(err)
+			return "", microerror.Mask(err)
 		}
 	case "kubernetes":
 		version, err = getLatestGithubRelease("kubernetes", "kubernetes")
 		// strip the "Kubernetes " prefix from the version
 		version, _ = strings.CutPrefix(version, "Kubernetes ")
 		if err != nil {
-			return componentVersion{}, microerror.Mask(err)
+			return "", microerror.Mask(err)
 		}
 	case "os-tooling":
 		version, err = getLatestGithubRelease("giantswarm", "capi-image-builder")
 		if err != nil {
-			return componentVersion{}, microerror.Mask(err)
+			return "", microerror.Mask(err)
 		}
 	default:
 		version, err = getLatestGithubRelease("giantswarm", name)
 		if err != nil {
-			return componentVersion{}, microerror.Mask(err)
+			return "", microerror.Mask(err)
 		}
 	}
 
 	version = strings.TrimPrefix(version, "v")
+
+	return version, nil
+}
+
+func findNewestComponent(name string) (componentVersion, error) {
+	version, err := findNewestComponentVersion(name)
+	if err != nil {
+		return componentVersion{}, microerror.Mask(err)
+	}
 
 	return componentVersion{
 		Version: version,